Fixing json issue that sent app into bad state
Do this by making the request again until json can be parsed. Hopefully won't lead to infinite loop

Some sort of json exception is happening intermittently- it seems the json can't be parsed

Unclear if this is a new issue because of local changes or if server has a regression. I see strange behavior and timeouts in mafia as well, so may not be me